+++ core/__init__.py
@@ -1,7 +1,7 @@
 from .cultivation_manager import CultivationManager
 from .equipment_manager import EquipmentManager
 from .breakthrough_manager import BreakthroughManager
-from .death_manager import DeathManager, DeathOutcome, SOUL_STATE_LEVEL_INDEX
+from .death_manager import DeathManager, DeathOutcome, SOUL_STATE_LEVEL_INDEX, resolve_death_config
 from .pill_manager import PillManager
 from .shop_manager import ShopManager
 from .storage_ring_manager import StorageRingManager
@@ -14,6 +14,7 @@ __all__ = [
     "DeathManager",
     "DeathOutcome",
     "SOUL_STATE_LEVEL_INDEX",
+    "resolve_death_config",
     "PillManager",
     "ShopManager",
     "StorageRingManager",